requirements.txt
MarkupSafe~=2.0.1
Werkzeug~=2.0.1
Jinja2~=3.0.1
click~=8.0.1
python-dotenv~=0.19.0
itsdangerous~=2.0.1
visitor~=0.1.3
certifi~=2021.5.30
urllib3~=1.26.6
WTForms~=2.3.3
SQLAlchemy~=1.4.23
dominate~=2.6.0
idna~=3.2
Flask~=2.0.1
requests~=2.26.0